melakukan migrate dan seeder ulang

+++ database/migrations/2023_11_14_182131_create_products_table.php
@@ -23,7 +23,7 @@ return new class extends Migration
             $table->integer('created_by')->nullable();
             $table->integer('updated_by')->nullable();
             $table->text('description')->nullable();
-            $table->decimal('price',15,2)->default(0);
+            $table->integer('price')->default(0);
             $table->string('unit', 100)->default('pcs');
             $table->decimal('discount_amount',15,2)->nullable();
             $table->integer('stock')->default(0);
